+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Added
+
+- Added `hostname` constraint.
+
 ## [v1.18.6](https://github.com/allenai/beaker-py/releases/tag/v1.18.6) - 2023-06-02
 
 ### Fixed